added extra class for pennylane fidelity kernel
